+++ style/src/checkbox.rs
@@ -6,8 +6,8 @@ use iced_core::{Background, Color};
 pub struct Style {
     pub background: Background,
     pub checkmark_color: Color,
-    pub border_radius: u16,
-    pub border_width: u16,
+    pub border_radius: f32,
+    pub border_width: f32,
     pub border_color: Color,
 }
 
@@ -25,8 +25,8 @@ impl StyleSheet for Default {
         Style {
             background: Background::Color(Color::from_rgb(0.95, 0.95, 0.95)),
             checkmark_color: Color::from_rgb(0.3, 0.3, 0.3),
-            border_radius: 5,
-            border_width: 1,
+            border_radius: 5.0,
+            border_width: 1.0,
             border_color: Color::from_rgb(0.6, 0.6, 0.6),
         }
     }
