+++ style/src/menu.rs
@@ -5,7 +5,7 @@ use iced_core::{Background, Color};
 pub struct Style {
     pub text_color: Color,
     pub background: Background,
-    pub border_width: u16,
+    pub border_width: f32,
     pub border_color: Color,
     pub selected_text_color: Color,
     pub selected_background: Background,
@@ -16,7 +16,7 @@ impl std::default::Default for Style {
         Self {
             text_color: Color::BLACK,
             background: Background::Color([0.87, 0.87, 0.87].into()),
-            border_width: 1,
+            border_width: 1.0,
             border_color: [0.7, 0.7, 0.7].into(),
             selected_text_color: Color::WHITE,
             selected_background: Background::Color([0.4, 0.4, 1.0].into()),
